Balance update 0.0.4
General
    -Damage dealt to an enemy cannot be higher than their current health (this affects how much ults are recharged when overkilling an enemy)
    -Cooldowns are no longer reduced in the lobby

Fish Ogre
    -Health and health per level reduced from 300/280 to 230/200
    -Passive healing reduced from 16% to 14%
    -Damage multiplier per level increased from 25% to 35%

    Primary Fire
        -Cooldown increased from 2.5 seconds to 3.5 seconds
        -Spread angle reduced from 60 degrees to 50 degrees

    Secondary Fire
        -Health removed reduced from 60% to 40%
        -Health to damage reduced from 100% to 50%
        -Gaunlet movement speed reduced from 560 to 460
        -Gaunlet now roots instead of silencing (duration reduced from 3.5 seconds to 2 seconds)

    Alt Ability
        -Lifesteal reduced from 50% to 25%

    Ultimate
        -No longer instantly heals full health
        -Killing during it no longer resets your gaunlet cooldown
        -Now transforms all incoming damage into heal for its duration (enemies don't get ultimate charge)
        -Duration reduced from 6.66 seconds to 4.5 seconds
        -Recharge speed (from both damage and time) reduced by 20%
Blondie
    -Health and health per level reduced from 220/200 to 180/160
    -Passive healing reduced from 12% to 11%
    -Damage multiplier per level increased from 28% to 30%

    Primary Fire
        -Now applies either silence or slow randomly (slow lasts 2 seconds while silence lasts 1.5)

Red Demon
    -Movement speed reduced from 200 to 180
    -Health reduced from 150 to 140
    -Damage multiplier per level increased from 33% to 35%

    Secondary Fire
        -Root duration increased from 2.5 seconds to 3.5 seconds

+++ include/buffs/storm_buff.hpp
@@ -10,12 +10,12 @@ public:
     virtual void onUpdate(sf::Time eTime);
 
     //these parameters are set by game mode when creating the buff
-    void setDamagePerSecond(float dmgPerSecond);
+    void setDamagePerSecond(u16 dmgPerSecond);
     void setMultiplierTimeTreshold(sf::Time multiplierTimeTreshold);
     void setDamageMultiplier(float dmgMultiplier);
 
 private:
-    float m_damagePerSecond;
+    u16 m_damagePerSecond;
     sf::Time m_multiplierTimeTreshold;
     float m_damageMultiplier;
